requirements.txt
annotated-types==0.5.0
attrs==23.1.0
blis==0.7.11
cassandra-driver==3.28.0
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
confection==0.1.3
cymem==2.0.8
exceptiongroup==1.1.3
fake-useragent==1.3.0
geomet==0.2.1.post1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.10
numpy==1.26.0
outcome==1.2.0
packaging==23.2
pandas==2.1.1
pathy==0.10.2
preshed==3.0.9
pydantic==2.4.2
pydantic_core==2.10.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.13.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.7.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typer==0.7.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wasabi==1.1.2
weasel==0.3.1
wsproto==1.2.0
